renamed launcher files
